DCP2-483 - Set up RabbitMQ and Sneakers
This introduces a minimal configuration for using RabbitMQ for
messaging/queueing and Sneakers for workers/background jobs, in
replacement of Redis and Resque, respectively. It does not, however,
remove either of the latter yet.

The built-in ActiveJob adapter for Sneakers (JobWrapper) only works on
the "default" queue and does not have an extension point. The existing
jobs target the "index" or "delete" queues, so we add our own Sneakers
workers that hook into ActiveJob and listen to a given queue. The base
worker is effectively the same as the JobWrapper, which we could extend,
but in an obvious location with a convenient name and apparent
implementation. We would likely have to adjust implementation if
JobWrapper changed anyway, so making it obvious that the adapter does
very little (decoding the event, dispatching the job, then acknowledging
it) is a net positive.

By placing the concrete listeners under app/workers, it is clear that
they are Rails-oriented and they are picked up by autoloading. This
makes the invocation via the sneakers:run rake concise (and similar to
resque:work):

  bin/rake sneakers:run WORKERS=IndexWorker,DeleteWorker