DCP2-483 - Wrap automation around the ingest jobs
This sequences the three main jobs (indexing, html generation, pdf
generation) to consider a finding aid fully ingested/processed by using
a fourth job for orchestration. It also adds a rake task to "ingest"
everything rather than simply "reindex".

The event names/shapes/contracts are completely unchecked. This is
stringly-typed, hash-oriented programming to add basic automation.
The simplest thing that could possibly work, you might say.

There is some very strange stuff going on with the requires and
autoloading of the lib directory. The setup with mixed require and
require_dependency is certainly not correct, but it is working.